requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg==3.2.2
psycopg-binary==3.2.2
PyJWT==2.9.0
python-dotenv==1.0.1
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.4